+++ main/src/test/java/net/oneandone/stool/setup/LibTest.java
@@ -79,5 +79,6 @@ public class LibTest {
         assertEquals("{\\+} \\u@\\h \\w \\$ ", stool.prompt);
         assertEquals(null, stool.defaults.get("https://svn.1and1.org/svn/controlpanel_app/controlpanel/").get("tomcat.perm"));
         assertEquals(Strings.toList(StageConfiguration.NOTIFY_OWNER), stage.notify);
+        assertEquals(0, stage.quota);
     }
 }